consolidate home page components
