+++ ohmg/frontend/svelte_components/src/components/home/LatestBlogPosts.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Link from '../../base/Link.svelte';
+  import Link from '../base/Link.svelte';
 
   $: blogItems = [];
